+++ rules/aws_resource_missing_tags_test.go
@@ -286,7 +286,7 @@ resource "aws_instance" "ec2_instance" {
 }
 
 resource "aws_instance" "ec2_instance_alias" {
-  provider = "aws.foo"
+  provider = aws.foo
   instance_type = "t2.micro"
 }`,
 			Config: `
@@ -376,6 +376,38 @@ rule "aws_resource_missing_tags" {
 						Filename: "module.tf",
 						Start:    hcl.Pos{Line: 10, Column: 1},
 						End:      hcl.Pos{Line: 10, Column: 44},
+					},
+				},
+			},
+		},
+		{
+			Name: "Provider reference no existent",
+			Content: `provider "aws" {
+  alias = "zoom"
+  default_tags {
+    tags = {
+      "Foo": "Bar"
+    }
+  }
+}
+
+resource "aws_instance" "ec2_instance" {
+  provider = aws.west
+  instance_type = "t2.micro"
+}`,
+			Config: `
+rule "aws_resource_missing_tags" {
+  enabled = true
+  tags = ["Foo"]
+}`,
+			Expected: helper.Issues{
+				{
+					Rule:    NewAwsResourceMissingTagsRule(),
+					Message: "The aws provider with alias \"west\" doesn't exist.",
+					Range: hcl.Range{
+						Filename: "module.tf",
+						Start:    hcl.Pos{Line: 11, Column: 17},
+						End:      hcl.Pos{Line: 11, Column: 22},
 					},
 				},
 			},