+++ rules/models/aws_appautoscaling_target_invalid_scalable_dimension.go
@@ -45,6 +45,7 @@ func NewAwsAppautoscalingTargetInvalidScalableDimensionRule() *AwsAppautoscaling
 			"elasticache:replication-group:NodeGroups",
 			"elasticache:replication-group:Replicas",
 			"neptune:cluster:ReadReplicaCount",
+			"sagemaker:variant:DesiredProvisionedConcurrency",
 		},
 	}
 }