+++ rules/models/aws_glue_ml_transform_invalid_worker_type.go
@@ -29,6 +29,8 @@ func NewAwsGlueMlTransformInvalidWorkerTypeRule() *AwsGlueMlTransformInvalidWork
 			"G.1X",
 			"G.2X",
 			"G.025X",
+			"G.4X",
+			"G.8X",
 		},
 	}
 }