+++ rules/models/aws_instance_invalid_type.go
@@ -668,6 +668,17 @@ func NewAwsInstanceInvalidTypeRule() *AwsInstanceInvalidTypeRule {
 			"m6idn.metal",
 			"r6in.metal",
 			"r6idn.metal",
+			"inf2.xlarge",
+			"inf2.8xlarge",
+			"inf2.24xlarge",
+			"inf2.48xlarge",
+			"trn1n.32xlarge",
+			"i4g.large",
+			"i4g.xlarge",
+			"i4g.2xlarge",
+			"i4g.4xlarge",
+			"i4g.8xlarge",
+			"i4g.16xlarge",
 		},
 	}
 }